Add router screens, styling clean up, rearrange components
