+++ src/components/cards/CardGrid.vue
@@ -97,5 +97,7 @@ function cardOpened(index: number) {
   flex-wrap: wrap;
   justify-content: space-evenly;
   height: 82%; /* This height also needs to be calculated */
+  width: 100%;
+  max-width: 350px;
 }
 </style>
