+++ src/globalStyles.css
@@ -44,15 +44,20 @@ body {
 }
 /*COLOR PALETTE*/
 .correct {
-  background-color: #6aaa64;
+  background-color: #6aaa64 !important;
 }
 
 .present {
-  background-color: #c9b458;
+  background-color: #c9b458 !important;
 }
 
 .absent {
-  background-color: #787c7e;
+  background-color: #787c7e !important;
+}
+.correct,
+.present,
+.absent {
+  color: white !important;
 }
 :root {
   --green: #6aaa64;
