--- src/views/WordleGame.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="gameWrapper">
-    <p>Wordle</p>
-  </div>
-</template>
