+++ src/views/WordleScreen.vue
@@ -1,0 +1,6 @@
+<script setup lang="ts">
+import WordleGame from '../components/wordle/WordleGame.vue';
+</script>
+<template>
+  <WordleGame />
+</template>
